backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-multipart==0.0.6
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pandas==2.1.3
openpyxl==3.1.2
celery==5.3.4
redis==5.0.1
python-dotenv==1.0.0
cryptography==41.0.8
jinja2==3.1.2
alembic==1.12.1
psycopg2-binary==2.9.9